+++ bindings/matrix-sdk-ffi/src/room_list.rs
@@ -15,10 +15,10 @@ use matrix_sdk_common::{SendOutsideWasm, SyncOutsideWasm};
 use matrix_sdk_ui::{
     room_list_service::filters::{
         new_filter_all, new_filter_any, new_filter_category, new_filter_deduplicate_versions,
-        new_filter_favourite, new_filter_fuzzy_match_room_name, new_filter_invite,
-        new_filter_joined, new_filter_low_priority, new_filter_non_left, new_filter_none,
-        new_filter_normalized_match_room_name, new_filter_not, new_filter_space, new_filter_unread,
-        BoxedFilterFn, RoomCategory,
+        new_filter_favourite, new_filter_fuzzy_match_room_name, new_filter_identifiers,
+        new_filter_invite, new_filter_joined, new_filter_low_priority, new_filter_non_left,
+        new_filter_none, new_filter_normalized_match_room_name, new_filter_not, new_filter_space,
+        new_filter_unread, BoxedFilterFn, RoomCategory,
     },
     unable_to_decrypt_hook::UtdHookManager,
 };
@@ -480,6 +480,7 @@ impl RoomListDynamicEntriesController {
 pub enum RoomListEntriesDynamicFilterKind {
     All { filters: Vec<RoomListEntriesDynamicFilterKind> },
     Any { filters: Vec<RoomListEntriesDynamicFilterKind> },
+    Identifiers { identifiers: Vec<String> },
     NonSpace,
     Space,
     NonLeft,
@@ -490,6 +491,7 @@ pub enum RoomListEntriesDynamicFilterKind {
     Favourite,
     LowPriority,
     NonLowPriority,
+    NonFavorite,
     Invite,
     Category { expect: RoomListFilterCategory },
     None,
@@ -525,6 +527,9 @@ impl From<RoomListEntriesDynamicFilterKind> for BoxedFilterFn {
             Kind::Any { filters } => Box::new(new_filter_any(
                 filters.into_iter().map(|filter| BoxedFilterFn::from(filter)).collect(),
             )),
+            Kind::Identifiers { identifiers } => Box::new(new_filter_identifiers(
+                identifiers.into_iter().map(|id| RoomId::parse(id).unwrap()).collect(),
+            )),
             Kind::NonSpace => Box::new(new_filter_not(Box::new(new_filter_space()))),
             Kind::Space => Box::new(new_filter_space()),
             Kind::NonLeft => Box::new(new_filter_non_left()),
@@ -533,6 +538,7 @@ impl From<RoomListEntriesDynamicFilterKind> for BoxedFilterFn {
             Kind::Favourite => Box::new(new_filter_favourite()),
             Kind::LowPriority => Box::new(new_filter_low_priority()),
             Kind::NonLowPriority => Box::new(new_filter_not(Box::new(new_filter_low_priority()))),
+            Kind::NonFavorite => Box::new(new_filter_not(Box::new(new_filter_favourite()))),
             Kind::Invite => Box::new(new_filter_invite()),
             Kind::Category { expect } => Box::new(new_filter_category(expect.into())),
             Kind::None => Box::new(new_filter_none()),